requirement.txt
MySQL-python==1.2.5
Django==1.6.4
South==0.8.4
Unipath==1.0
django-extensions
django-rest-swagger
djangorestframework==2.3.13
ipdb
ipython
pyasn1==0.1.7
pycrypto==2.6.1
python-keyczar==0.71c
six==1.5.2
wsgiref==0.1.2
Werkzeug
fake-factory
django-cors-headers==0.12
pytz==2014.3
